// src/types/types.tsx
export interface QuestionPropsInterface {
    qid: string;
    title: string;
    text: string;
    tagIds: string[];
    askedBy: string;
    askDate: Date;
    ansIds: string[];
    views: number;
    getAnswerCount: () => number;
    addAnswer: (aid: string) => void;
    getAnswersId: () => string[];
    getTagsId: () => string[];
    calculateTimeElapsed: () => string;
    getQuestionViews: () => number;
    addViewCount: () => void;
    setNewestAnswerDate: (date: Date) => void;
}

/**
 * An interface representing the input parameters
 * the constructor of the Question class expects
 */
export interface QuestionParamType {
    qid: string;
    title: string;
    text: string;
    tagIds: string[];
    askedBy: string;
    askDate: Date;
    ansIds: string[];
    views: number;
}